Fix syntax error with lack of else and a logic bug
There was a missing `else` clause.

Also, in the event that the invoke-webrequest fails to retrieve a photo from zenifits the script will default to using the photo from the previous person.  This logic error has also been fixed here by ensuring that each loop starts with a null variable.

Finally, fixed a debug line to output the proper URL of the photo being fetched.